+++ foxes/input/yaml/__init__.py
@@ -1,3 +1,3 @@
-from .dict import read_dict, run_dict
+from .dict import read_dict, run_dict, run_outputs
 from .yaml import foxes_yaml
 from .windio.windio import read_windio, foxes_windio