+++ foxes/output/__init__.py
@@ -6,7 +6,7 @@ from .round import round_defaults
 from .output import Output
 from .farm_layout import FarmLayoutOutput
 from .farm_results_eval import FarmResultsEval
-from .rose_plot import RosePlotOutput, StatesRosePlotOutput
+from .rose_plot import RosePlotOutput, StatesRosePlotOutput, WindRoseBinPlot
 from .results_writer import ResultsWriter
 from .state_turbine_map import StateTurbineMap
 from .turbine_type_curves import TurbineTypeCurves